feat: list all analysis and list analysis by id

+++ app/src/main/java/com/banap/banap/core/ui/util/fertilizerCalculation.kt
@@ -83,9 +83,9 @@ object FertilizerCalculator {
         }
 
         return NPKResult(
-            nitrogen = nitrogen,
-            phosphor = phosphorValue,
-            potassium = potassiumValue
+            nitrogen = nitrogen.toDouble(),
+            phosphor = phosphorValue.toDouble(),
+            potassium = potassiumValue.toDouble()
         )
     }
