+++ app/src/main/java/com/banap/banap/domain/model/analysis/NPKResult.kt
@@ -1,7 +1,7 @@
 package com.banap.banap.domain.model.analysis
 
 data class NPKResult(
-    val nitrogen: Int,
-    val phosphor: Int,
-    val potassium: Int
+    val nitrogen: Double,
+    val phosphor: Double,
+    val potassium: Double
 )
